requirements.txt
asgiref==3.4.1
Django==3.2.6
django-simple-history==3.0.0
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
